+++ radio/src/gui/480x272/layout.h
@@ -24,6 +24,7 @@
 #include <list>
 #include "widgets_container.h"
 #include "lcd.h"
+#include "zone.h"
 
 #define MAX_LAYOUT_ZONES               10
 #define MAX_LAYOUT_OPTIONS             10
