+++ radio/src/gui/480x272/theme.h
@@ -22,6 +22,7 @@
 #define _THEME_H_
 
 #include <list>
+#include "zone.h"
 
 class BitmapBuffer;
