+++ radio/src/gui/480x272/topbar.h
@@ -22,6 +22,7 @@
 #define _TOPBAR_H_
 
 #include "widgets_container.h"
+#include "zone.h"
 
 #define MAX_TOPBAR_ZONES               4
 #define MAX_TOPBAR_OPTIONS             1 // just because of VC++ which doesn't like 0-size arrays :(
